+++ Assignment.Interfaces/IAddress.cs
@@ -1,0 +1,9 @@
+﻿namespace Assignment;
+
+public interface IAddress
+{
+    string StreetAddress { get; }
+    string City { get; }
+    string State { get; }
+    string Zip { get; }
+}
