--- Assignment/Assignment.Interfaces/IAddress.cs
@@ -1,9 +1,0 @@
-﻿namespace Assignment;
-
-public interface IAddress
-{
-    string StreetAddress { get; }
-    string City { get; }
-    string State { get; }
-    string Zip { get; }
-}
