--- Assignment/Assignment.Interfaces/IPerson.cs
@@ -1,10 +1,0 @@
-﻿namespace Assignment;
-
-public interface IPerson
-{
-    string FirstName { get; }
-    string LastName { get; }
-    IAddress Address { get; }
-
-    string EmailAddress { get; }
-}
