--- Assignment/Directory.Build.props
@@ -1,10 +1,0 @@
-<Project>
-<PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <EnableNETAnalyzers>true</EnableNETAnalyzers>
-    <CodeAnalysisTreatWarningsAsErrors>true</CodeAnalysisTreatWarningsAsErrors>
-	<LangVersion>latest</LangVersion>
-  </PropertyGroup>
-</Project>
